requirements.txt
Django==4.1.5
djangorestframework==3.14.0
django-cors-headers==3.13.0
djangorestframework-simplejwt~=5.2.2
djangorestframework-camel-case==1.3.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
django-debug-toolbar==3.8.1
